cleanup(weight-transfer): clarify CUDA memory sizing
